@@ -1,9 +1,4 @@
-//
-// Created by Nidhogg on 2024/5/28.
-//
-
 #include "http_connect.h"
-
 #include <mysql/mysql.h>
 #include <mutex>
 #include <fstream>
@@ -21,28 +16,47 @@ const char *error_500_form = "There was an unusual problem serving the request f
 
 std::mutex m_lock;
 map<string, string> users;
+int http_conn::m_user_count = 0;
+int http_conn::m_epollfd = -1;
 //TODO::初始化连接池
 void http_conn::initmysql_result(connection_pool *connPool)
 {
     //先从连接池中取一个连接
+    MYSQL* mysql = NULL;
+    connectionRAII mysqlcon(&mysql, connPool);
 
+    
+    //在user表中检索username，passwd数据，浏览器端输入
+    if (mysql_query(mysql, "SELECT username,passwd FROM user"))
+    {
+        LOG_ERROR("SELECT error:%s\n", mysql_error(mysql));
+    }
 
     //从表中检索完整的结果集
-
-
-    //返回结果集中的列数
-
+    MYSQL_RES *result = mysql_store_result(mysql);
 
     //返回所有字段结构的数组
-
-
+    MYSQL_FIELD *fields = mysql_fetch_fields(result);
     //从结果集中获取下一行，将对应的用户名和密码，存入map中
+    while (MYSQL_ROW row = mysql_fetch_row(result))
+    {
+        string temp1(row[0]);
+        string temp2(row[1]);
+        users[temp1] = temp2;
+    }
+}
+void http_conn::pick_one_sql_connect(connection_pool* connPool)
+{
+    connectionRAII mysqlcon(&this->mysql,connPool);
 }
 
 //TODO::对文件描述符设置非阻塞
 int setnonblocking(int fd)
 {
-
+    int old_option = fcntl(fd, F_GETFL);
+    int new_option = old_option | O_NONBLOCK;
+    fcntl(fd, F_SETFL, new_option);
+    return old_option;
 }
 
 //TODO::将内核事件表注册读事件，ET模式，选择开启EPOLLONESHOT
@@ -65,21 +79,35 @@ void addfd(int epollfd, int fd, bool one_shot, int TRIGMode)
 //TODO::从内核时间表删除描述符
 void removefd(int epollfd, int fd)
 {
-
+    epoll_ctl(epollfd, EPOLL_CTL_DEL, fd, 0);
+    close(fd);
 }
 
 //TODO::将事件重置为EPOLLONESHOT
 void modfd(int epollfd, int fd, int ev, int TRIGMode)
 {
+    epoll_event event;
+    event.data.fd = fd;
 
+    if (1 == TRIGMode)
+        event.events = ev | EPOLLET | EPOLLONESHOT | EPOLLRDHUP;
+    else
+        event.events = ev | EPOLLONESHOT | EPOLLRDHUP;
+
+    epoll_ctl(epollfd, EPOLL_CTL_MOD, fd, &event);
 }
 
-int http_conn::m_user_count = 0;
-int http_conn::m_epollfd = -1;
 
 //TODO::关闭连接，关闭一个连接，客户总量减一
 void http_conn::close_conn(bool real_close)
 {
+    if (real_close && (m_sockfd != -1))
+    {
+        printf("close %d\n", m_sockfd);
+        removefd(m_epollfd, m_sockfd);
+        m_sockfd = -1;
+        m_user_count--;
+    }
 
 }
 
@@ -88,19 +116,19 @@ void http_conn::init(int sockfd, const sockaddr_in &addr, char *root, int TRIGMo
                      int close_log, string user, string passwd, string sqlname)
 {
     this->m_sockfd = sockfd;
-    this->m_address = addr;
+    this->m_address = addr; 
     addfd(m_epollfd, sockfd, true, m_TRIGMode);
-    http_conn::m_user_count++;
+    http_conn::m_user_count++;  
     //当浏览器出现连接重置时，可能是网站根目录出错或http响应格式出错或者访问的文件中内容完全为空
     this->doc_root = root;
     this->m_TRIGMode = TRIGMode;
     this->m_close_log = close_log;
+    
 
     strcpy(this->sql_user, user.c_str());
     strcpy(this->sql_passwd, passwd.c_str());
     strcpy(this->sql_name, sqlname.c_str());
-    this->init();
-    
+    this->init();    
 }
 
 //TODO::初始化新接受的连接
@@ -207,28 +235,146 @@ bool http_conn::read_once()
         }
         return true;
     }
-    return false;
-    
+    return false;    
 }
 
 //TODO::解析http请求行，获得请求方法，目标url及http版本号
 http_conn::HTTP_CODE http_conn::parse_request_line(char *text)
 {
+    //在HTTP报文中，请求行用来说明请求类型
+    //要访问的资源以及所使用的HTTP版本，其中各个部分之间通过\t或空格分隔。
+    //请求行中最先含有空格和\t任一字符的位置并返回
 
-    //当url为/时，显示判断界面
+    //strpbrk在源字符串（s1）中找出最先含有搜索字符串（s2）
+    //中任一字符的位置并返回，若找不到则返回空指针
+    m_url = strpbrk(text, " \t");
+    //如果没有空格或\t，则报文格式有误
+    if (!m_url)
+    {
+        return HTTP_CODE::BAD_REQUEST;
+    }
+    //将该位置改为\0，用于将前面数据取出
+    //(已读取的数据不再会匹配到\t）
+    *m_url++ = '\0';
 
+    //取出数据，并通过与GET和POST比较，以确定请求方式
+    char *method = text;
+    if (strcasecmp(method, "GET") == 0)
+        m_method = METHOD::GET;
+    else if (strcasecmp(method, "POST") == 0)
+    {
+        m_method = METHOD::POST;
+        cgi = 1;
+    }
+    else
+        return HTTP_CODE::BAD_REQUEST;
+    //m_url此时跳过了第一个空格或\t字符，但不知道之后是否还有
+    //将m_url向后偏移，通过查找
+    //继续跳过空格和\t字符，指向请求资源的第一个字符
+
+    //strspn:检索字符串 str1 中第一个不在字符串 str2 中出现的字符下标
+    //即跳过匹配的字符串片段
+    m_url += strspn(m_url, " \t");
+    //相同逻辑，判断HTTP版本号
+    m_version = strpbrk(m_url, " \t");
+    if (!m_version)
+        return HTTP_CODE::BAD_REQUEST;
+    *m_version++ = '\0';
+    m_version += strspn(m_version, " \t");
+    //仅支持HTTP/1.1
+    if (strcasecmp(m_version, "HTTP/1.1") != 0)
+        return HTTP_CODE::BAD_REQUEST;
+    
+    //对请求资源前7个字符进行判断
+    //这里主要是有些报文的请求资源中会带有http://
+    //这里需要对这种情况进行单独处理
+    if (strncasecmp(m_url, "http://", 7) == 0)
+    {
+        m_url += 7;
+        m_url = strchr(m_url, '/');
+    }
+
+    //同样增加https情况
+    if (strncasecmp(m_url, "https://", 8) == 0)
+    {
+        m_url += 8;
+        m_url = strchr(m_url, '/');
+    }
+
+    //一般的不会带有上述两种符号，直接是单独的/或/后面带访问资源
+    if (!m_url || m_url[0] != '/')
+        return HTTP_CODE::BAD_REQUEST;
+    //当url为/时，显示欢迎界面
+    if (strlen(m_url) == 1)
+        strcat(m_url, "judge.html");
+    
+    //请求行处理完毕，将主状态机转移处理请求头
+    m_check_state = CHECK_STATE::CHECK_STATE_HEADER;
+    return HTTP_CODE::NO_REQUEST;
 }
 
 //TODO::解析http请求的一个头部信息
 http_conn::HTTP_CODE http_conn::parse_headers(char *text)
 {
+    //判断是空行还是请求头
+    if (text[0] == '\0')
+    {
+        //判断是GET还是POST请求
+        //!0 is POST
+        if (m_content_length != 0)
+        {
+            //POST需要跳转到消息体处理状态
+            m_check_state = CHECK_STATE::CHECK_STATE_CONTENT;
+            return HTTP_CODE::NO_REQUEST;
+        }
+        //==0 is GET
+        return HTTP_CODE::GET_REQUEST;
+    }
+    //解析请求头部connection字段
+    else if (strncasecmp(text, "Connection:", 11) == 0)
+    {
+        text += 11;
 
+        //跳过空格和\t字符
+        text += strspn(text, " \t");
+        if (strcasecmp(text, "keep-alive") == 0)
+        {
+            //如果是长连接，则将linger标志设置为true
+            m_linger = true;
+        }
+    }
+    //解析请求头部Content-length字段
+    else if (strncasecmp(text, "Content-length:", 15) == 0)
+    {
+        text += 15;
+        text += strspn(text, " \t");
+        m_content_length = atol(text);
+    }
+    //解析请求头部Host字段
+    else if (strncasecmp(text, "Host:", 5) == 0)
+    {
+        text += 5;
+        text += strspn(text, " \t");
+        m_host = text;
+    }
+    else
+    {
+        LOG_INFO("oop!unknow header: %s", text);
+    }
+    return HTTP_CODE::NO_REQUEST;
 }
 
 //TODO::判断http请求是否被完整读入
 http_conn::HTTP_CODE http_conn::parse_content(char *text)
 {
-
+    if (m_read_idx >= (m_content_length + m_checked_idx))
+    {
+        text[m_content_length] = '\0';
+        //POST请求中最后为输入的用户名和密码
+        m_string = text;
+        return HTTP_CODE::GET_REQUEST;
+    }
+    return HTTP_CODE::NO_REQUEST;
 }
 
 //TODO::有限状态机处理请求报文
@@ -242,7 +388,7 @@ http_conn::HTTP_CODE http_conn::process_read()
     {
         text = get_line();
         m_start_line = m_checked_idx;
-        //LOG_INFO("%s", text);
+        LOG_INFO("%s", text);
         switch (m_check_state)
         {
         case CHECK_STATE::CHECK_STATE_REQUESTLINE:
@@ -276,7 +422,6 @@ http_conn::HTTP_CODE http_conn::process_read()
         }
     }
     return HTTP_CODE::NO_REQUEST;
-
 }
 
 //TODO::生成响应
@@ -284,7 +429,7 @@ http_conn::HTTP_CODE http_conn::do_request()
 {
     strcpy(m_real_file, doc_root);
     int len = strlen(doc_root);
-    //printf("m_url:%s\n", m_url);
+    printf("m_url:%s\n", m_url);
     const char *p = strrchr(m_url, '/');
 
     //处理cgi
@@ -327,11 +472,13 @@ http_conn::HTTP_CODE http_conn::do_request()
 
             if (users.find(name) == users.end())
             {
-                m_lock.lock();
+                //mysql = mysql_init(NULL);
+                //bool ok = false;
+                //ok = mysql_real_connect(mysql, "localhost", "root", "18861817269lwj", "webserDb", 3306, 0, NULL);                   
+                std::unique_lock<std::mutex> locker(m_lock);
                 int res = mysql_query(mysql, sql_insert);
                 users.insert(pair<string, string>(name, password));
-                m_lock.unlock();
-
+                locker.unlock();
                 if (!res)
                     strcpy(m_url, "/log.html");
                 else
@@ -356,7 +503,6 @@ http_conn::HTTP_CODE http_conn::do_request()
         char *m_url_real = (char *)malloc(sizeof(char) * 200);
         strcpy(m_url_real, "/register.html");
         strncpy(m_real_file + len, m_url_real, strlen(m_url_real));
-
         free(m_url_real);
     }
     //如果请求资源为/1，表示跳转登录界面
@@ -374,7 +520,6 @@ http_conn::HTTP_CODE http_conn::do_request()
         char *m_url_real = (char *)malloc(sizeof(char) * 200);
         strcpy(m_url_real, "/picture.html");
         strncpy(m_real_file + len, m_url_real, strlen(m_url_real));
-
         free(m_url_real);
     }
     //如果请求资源为/6，表示跳转video
@@ -383,7 +528,6 @@ http_conn::HTTP_CODE http_conn::do_request()
         char *m_url_real = (char *)malloc(sizeof(char) * 200);
         strcpy(m_url_real, "/video.html");
         strncpy(m_real_file + len, m_url_real, strlen(m_url_real));
-
         free(m_url_real);
     }
     //如果请求资源为/7，表示跳转weixin
@@ -395,13 +539,18 @@ http_conn::HTTP_CODE http_conn::do_request()
         free(m_url_real);
     }
     else
+        strncpy(m_real_file + len, m_url, FILENAME_LEN - len - 1);
         //如果以上均不符合，即不是登录和注册，直接将url与网站目录拼接
         //这里的情况是welcome界面，请求服务器上的一个图片
-        strncpy(m_real_file + len, m_url, FILENAME_LEN - len - 1);
     //通过stat获取请求资源文件信息，成功则将信息更新到m_file_stat结构体
     //失败返回NO_RESOURCE状态，表示资源不存在
     if (stat(m_real_file, &m_file_stat) < 0)
+    {
+        std::cout << "找不到该资源" << std::endl;
         return HTTP_CODE::NO_RESOURCE;
+
+    }
+        
 
     //判断文件的权限，是否可读，不可读则返回FORBIDDEN_REQUEST状态
     if (!(m_file_stat.st_mode & S_IROTH))
@@ -416,97 +565,230 @@ http_conn::HTTP_CODE http_conn::do_request()
     m_file_address = (char *)mmap(0, m_file_stat.st_size, PROT_READ, MAP_PRIVATE, fd, 0);
     //避免文件描述符的浪费和占用
     close(fd);
-
     //表示请求文件存在，且可以访问
     return HTTP_CODE::FILE_REQUEST;
-
-    //处理cgi
-
-        //根据标志判断是登录检测还是注册检测
-
-
-        //将用户名和密码提取出来
-        //user=123&passwd=123
-
-            //如果是注册，先检测数据库中是否有重名的
-            //没有重名的，进行增加数据
-
-            //如果是登录，直接判断
-            //若浏览器端输入的用户名和密码在表中可以查找到，返回1，否则返回0
-
-
 }
 //TODO::将内存映射区释放
 void http_conn::unmap()
 {
-
+    if (m_file_address)
+    {
+        munmap(m_file_address, m_file_stat.st_size);
+        m_file_address = 0;
+    }
 }
 //TODO::写http响应
 bool http_conn::write()
 {
+    int temp = 0;
+    //表示响应报文为空，一般不会出现这种情况
+    if (bytes_to_send == 0)
+    {
+        modfd(m_epollfd, m_sockfd, EPOLLIN, m_TRIGMode);
+        init();
+        return true;
+    }
 
+    while (1)
+    {
+        //将响应报文的状态行、消息头、空行和响应正文发送给浏览器端
+        temp = writev(m_sockfd, m_iv, m_iv_count);
+        //error
+        if (temp < 0)
+        {
+            //判断缓冲区是否满了
+            if (errno == EAGAIN)
+            {
+                modfd(m_epollfd, m_sockfd, EPOLLOUT, m_TRIGMode);
+                return true;
+            }
+            unmap();
+            return false;
+        }
+        //更新已发送字节
+        bytes_have_send += temp;
+        //更新wei发送字节
+        bytes_to_send -= temp;
+        //第一个iovec头部信息的数据已发送完，发送第二个iovec数据
+        if (bytes_have_send >= m_iv[0].iov_len)
+        {
+            //不再继续发送头部信息
+            m_iv[0].iov_len = 0;
+            m_iv[1].iov_base = m_file_address + (bytes_have_send - m_write_idx);
+            m_iv[1].iov_len = bytes_to_send;
+        }
+        //继续发送第一个iovec头部信息的数据
+        else
+        {
+            m_iv[0].iov_base = m_write_buf + bytes_have_send;
+            m_iv[0].iov_len = m_iv[0].iov_len - bytes_have_send;
+        }
+
+        //判断条件，数据已全部发送完
+        if (bytes_to_send <= 0)
+        {
+            //如果发送失败，但不是缓冲区问题，取消映射
+            unmap();
+            //重新注册写事件
+            modfd(m_epollfd, m_sockfd, EPOLLIN, m_TRIGMode);
+
+            //浏览器的请求为长连接
+            if (m_linger)
+            {
+                //重新初始化HTTP对象**
+                init();
+                return true;
+            }
+            else
+            {
+                return false;
+            }
+        }
+    }
 }
 
 //TODO::往写缓冲中写入待发送的数据
 bool http_conn::add_response(const char *format, ...)
 {
-
+    //如果写入内容超出m_write_buf大小则报错
+    if (m_write_idx >= WRITE_BUFFER_SIZE)
+        return false;
+    //定义可变参数列表
+    va_list arg_list;
+    //将变量arg_list初始化为传入参数
+    va_start(arg_list, format);
+    //将数据format从可变参数列表写入缓冲区写，返回写入数据的长度
+    int len = vsnprintf(m_write_buf + m_write_idx, WRITE_BUFFER_SIZE - 1 - m_write_idx, format, arg_list);
+    //如果写入的数据长度超过缓冲区剩余空间，则报错
+    if (len >= (WRITE_BUFFER_SIZE - 1 - m_write_idx))
+    {
+        va_end(arg_list);
+        return false;
+    }
+    //更新m_write_idx位置
+    m_write_idx += len;
+    //清空可变参列表
+    va_end(arg_list);
+    //LOG_INFO("request:%s", m_write_buf);
+    return true;
 }
 
-
+//添加状态行
 bool http_conn::add_status_line(int status, const char *title)
 {
-
+    return add_response("%s %d %s\r\n", "HTTP/1.1", status, title);
 }
-
+//添加消息报头，具体的添加文本长度、连接状态和空行
 bool http_conn::add_headers(int content_len)
 {
-
+    return add_content_length(content_len) && add_linger() &&
+           add_blank_line();
 }
-
+//添加Content-Length，表示响应报文的长度
 bool http_conn::add_content_length(int content_len)
 {
-
+    return add_response("Content-Length:%d\r\n", content_len);
 }
-
+//添加文本类型，这里是html
 bool http_conn::add_content_type()
 {
-
+    return add_response("Content-Type:%s\r\n", "text/html");
 }
+//添加连接状态，通知浏览器端是保持连接还是关闭
 bool http_conn::add_linger()
 {
-
+    return add_response("Connection:%s\r\n", (m_linger == true) ? "keep-alive" : "close");
 }
+//添加空行
 bool http_conn::add_blank_line()
 {
-
+    return add_response("%s", "\r\n");
 }
+//添加文本content
 bool http_conn::add_content(const char *content)
 {
-
+    return add_response("%s", content);
 }
-//TODO::根据服务器处理HTTP请求的结果，决定返回给客户端的内容
+//TODO::根据服务器处理HTTP请求的结果，决定返回给客户端的内容，即生成响应报文
 bool http_conn::process_write(HTTP_CODE ret)
 {
-
+    switch (ret)
+    {
+    case HTTP_CODE::INTERNAL_ERROR:
+    {
+        add_status_line(500, error_500_title);
+        add_headers(strlen(error_500_form));
+        if (!add_content(error_500_form))
+            return false;
+        break;
+    }
+    case HTTP_CODE::BAD_REQUEST:
+    {
+        add_status_line(404, error_404_title);
+        add_headers(strlen(error_404_form));
+        if (!add_content(error_404_form))
+            return false;
+        break;
+    }
+    case HTTP_CODE::FORBIDDEN_REQUEST:
+    {
+        add_status_line(403, error_403_title);
+        add_headers(strlen(error_403_form));
+        if (!add_content(error_403_form))
+            return false;
+        break;
+    }
+    case HTTP_CODE::FILE_REQUEST:
+    {
+        add_status_line(200, ok_200_title);
+        if (m_file_stat.st_size != 0)
+        {
+            add_headers(m_file_stat.st_size);
+            m_iv[0].iov_base = m_write_buf;
+            m_iv[0].iov_len = m_write_idx;
+            m_iv[1].iov_base = m_file_address;
+            m_iv[1].iov_len = m_file_stat.st_size;
+            m_iv_count = 2;
+            bytes_to_send = m_write_idx + m_file_stat.st_size;
+            return true;
+        }
+        else
+        {
+            const char *ok_string = "<html><body></body></html>";
+            add_headers(strlen(ok_string));
+            if (!add_content(ok_string))
+                return false;
+        }
+    }
+    default:
+        return false;
+    }
+    m_iv[0].iov_base = m_write_buf;
+    m_iv[0].iov_len = m_write_idx;
+    m_iv_count = 1;
+    bytes_to_send = m_write_idx;
+    return true;
 }
 //TODO::由线程池中的工作线程调用，这是处理http请求的入口函数
-void http_conn::process(http_conn& user)
+void http_conn::process(http_conn* user, connection_pool* connPool)
 {
+    //数据库连接池取连接
+    user->pick_one_sql_connect(connPool);
+    std::cout << "处理http请求" <<std::endl;
     //NO_REQUEST，表示请求不完整，需要继续接收请求数据
-    HTTP_CODE read_ret = user.process_read();
-    if (read_ret == http_conn::HTTP_CODE::NO_REQUEST)
+    HTTP_CODE read_ret = user->process_read();
+    if (read_ret == HTTP_CODE::NO_REQUEST)
     {
         //注册并监听读事件
-        modfd(user.m_epollfd, user.m_sockfd, EPOLLIN, user.m_TRIGMode);
+        modfd(user->m_epollfd, user->m_sockfd, EPOLLIN, user->m_TRIGMode);
         return;
     }
     //调用process_write完成报文响应
-    bool write_ret = user.process_write(read_ret);
+    bool write_ret = user->process_write(read_ret);
     if (!write_ret)
     {
-        user.close_conn();
+        user->close_conn();
     }
     //注册并监听写事件
-    modfd(user.m_epollfd, user.m_sockfd, EPOLLOUT, user.m_TRIGMode);
+    modfd(user->m_epollfd, user->m_sockfd, EPOLLOUT, user->m_TRIGMode);
 }
