@@ -22,16 +22,21 @@
 #include <sys/uio.h>
 #include <map>
 #include <mysql/mysql.h>
-#include "../Imysql/sql_connection_pool.h"
+#include "../sql_connection_pool.h"
 #include "../timer/timer.h"
 #include "../log/log.h"
+#include "../threadpool/threadpool.h"
 
 class http_conn
 {
 public:
+    //读取文件长度上限
     static const int FILENAME_LEN = 200;
+    //读缓存大小
     static const int READ_BUFFER_SIZE = 2048;
+    //写缓存大小
     static const int WRITE_BUFFER_SIZE = 1024;
+    //HTTP方法名
     enum class METHOD
     {
         GET = 0,
@@ -44,7 +49,7 @@ public:
         CONNECT,
         PATH
     };
-    //主状态机的状态
+    //主状态机的状态,检查请求报文中元素
     enum class CHECK_STATE
     {
         CHECK_STATE_REQUESTLINE = 0,
@@ -52,6 +57,7 @@ public:
         CHECK_STATE_CONTENT
     };
     static CHECK_STATE check_state;
+    //HTTP状态码
     enum class HTTP_CODE
     {
         NO_REQUEST,
@@ -64,6 +70,7 @@ public:
         CLOSED_CONNECTION
     };
     static HTTP_CODE http_code;
+    //从状态机的状态，文本解析是否成功
     enum class LINE_STATUS
     {
         LINE_OK = 0,
@@ -77,31 +84,49 @@ public:
     ~http_conn() {}
 
 public:
+    //初始化套接字
     void init(int sockfd, const sockaddr_in &addr, char *, int, int, string user, string passwd, string sqlname);
+    //关闭HTTP连接
     void close_conn(bool real_close = true);
-    static void process(http_conn& user);
+    //http处理函数
+    static void process(http_conn* user, connection_pool* connPool);
+    //读取浏览器发送的数据
     bool read_once();
+    //给相应报文中写入数据
     bool write();
     sockaddr_in *get_address()
     {
         return &m_address;
     }
+    //初始化数据库读取线程
     void initmysql_result(connection_pool *connPool);
-    int timer_flag;
-    int improv;
+    //从数据库连接池取连接
+    void pick_one_sql_connect(connection_pool *connPool);
+    int timer_flag;//是否关闭连接
+    int improv;//是否正在处理数据中
 
 
 private:
     void init();
+    //从m_read_buf读取，并处理请求报文
     HTTP_CODE process_read();
+    //向m_write_buf写入响应报文数据
     bool process_write(HTTP_CODE ret);
+    //主状态机解析报文中的请求行数据
     HTTP_CODE parse_request_line(char *text);
+    //主状态机解析报文中的请求头数据
     HTTP_CODE parse_headers(char *text);
+    //主状态机解析报文中的请求内容
     HTTP_CODE parse_content(char *text);
+    //生成响应报文
     HTTP_CODE do_request();
+    //m_start_line是已经解析的字符
+	//get_line用于将指针向后偏移，指向未处理的字符
     char *get_line() { return m_read_buf + m_start_line; };
+    //从状态机读取一行，分析是请求报文的哪一部分
     LINE_STATUS parse_line();
     void unmap();
+    //根据响应报文格式，生成对应8个部分，以下函数均由do_request调用
     bool add_response(const char *format, ...);
     bool add_content(const char *content);
     bool add_status_line(int status, const char *title);
@@ -120,15 +145,24 @@ public:
 private:
     int m_sockfd;
     sockaddr_in m_address;
+    //存储读取的请求报文数据
     char m_read_buf[READ_BUFFER_SIZE];
     //缓冲区中m_read_buf中数据的最后一个字节的下一个位置
     long m_read_idx;
+    //m_read_buf读取的位置m_checked_idx
     long m_checked_idx;
+    //m_read_buf中已经解析的字符个数
     int m_start_line;
+    //存储发出的响应报文数据
     char m_write_buf[WRITE_BUFFER_SIZE];
+    //指示buffer中的长度
     int m_write_idx;
+    //主状态机的状态
     CHECK_STATE m_check_state;
+    //请求方法
     METHOD m_method;
+    //以下为解析请求报文中对应的6个变量
+    //存储读取文件的名称
     char m_real_file[FILENAME_LEN];
     char *m_url;
     char *m_version;
@@ -136,18 +170,22 @@ private:
     long m_content_length;
     bool m_linger;
     char *m_file_address;
+    //读取服务器上的文件地址
     struct stat m_file_stat;
-    struct iovec m_iv[2];
+    //io向量机制iovec
+    struct iovec m_iv[2];   
     int m_iv_count;
     int cgi;        //是否启用的POST
     char *m_string; //存储请求头数据
+    //剩余发送字节数
     int bytes_to_send;
+    //已发送字节数
     int bytes_have_send;
     char *doc_root;
 
     map<string, string> m_users;
-    int m_TRIGMode;
-    int m_close_log;
+    int m_TRIGMode;//触发模式
+    int m_close_log;//是否开启日志
 
     char sql_user[100];
     char sql_passwd[100];
